massive cleanup and incomplete mov of satellite code, now improving the linux emulator
